docs(#656): recapture every gallery light from the current corpus
The gallery previews were the visible half of #656 and none of them had been
regenerated since the datasets landed. 38 were captured on 19–21 July, before
the corpus moved onto historical data, and the 6 interaction lights predate
#702's data change — so a visitor still saw the monotonous default-theme
gallery the issue complains about, no matter what the examples actually render
now. point/scatter-color's card still read "Penguin flippers vs body mass".

All 44 are recaptured from the built docs site. Five had also drifted off the
canonical frame geometry (641×441, 800×760, 640×480, 640×401, 640×360) because
they were made by hand rather than by a capture script; every light is now
exactly the manifest's vrWidth × vrHeight.

capture-interaction-gallery-lights.ts becomes capture-gallery-lights.ts: the
example filter is an optional id-prefix argument instead of a hardcoded
`startsWith("interaction/")`, so `bun run gallery:previews:capture` covers the
corpus and `… interaction/` still covers the old subset. Two changes fall out of
generalizing it:

- Plot readiness is counted from the DOM rather than a per-example table. The
  docs site prerenders every <GGPlot>, so all `.gg-plot-root` elements are in
  the served HTML before hydration (verified: 3 for legend-focus, 2 for
  linked-views, 1 for scatter-color) — no table to go stale when an example
  gains a chart.
- Navigation and action budgets go to 120s. The 10k-mark canvas specimen paints
  on the main thread and outlasts Playwright's 30s default, which is why
  vr.spec.ts already grants that scenario the same. Readiness is still gated by
  settle(), so a slow example waits instead of shipping a half-drawn preview.

The geometry test now walks the whole corpus instead of the interaction subset —
it is the check that would have caught those five drifted captures.
