+++ package.json
@@ -26,7 +26,7 @@
     "lesson:charts:check": "bun scripts/gen-lesson-charts.ts --check",
     "gallery:previews:gen": "bun scripts/gen-gallery-previews.ts",
     "gallery:previews:check": "bun scripts/gen-gallery-previews.ts --check",
-    "gallery:previews:capture-interaction": "bun scripts/capture-interaction-gallery-lights.ts",
+    "gallery:previews:capture": "bun scripts/capture-gallery-lights.ts",
     "playground:seeds:gen": "bun scripts/gen-playground-seeds.ts",
     "playground:seeds:check": "bun scripts/gen-playground-seeds.ts --check",
     "check:docs-metadata": "bun scripts/check-docs-metadata.ts",
